MCCS-199 Remove failing BDD test in k8s deployment
The allocate BDD test fails when tested in a cluster.
It was an enable test that I tried to change to an allocate test,
so I have removed it. Interestingly though, the allocate BDD test
is commented out below it. This will be dealt with in my BDD
test implementation tasks.