Added skeleton tpm driver. Based on pyaavs Tile device (here HwTile) for TPM V1.2. At the moment supports just connection, firmware download, and some basic board monitoring (temperature, voltage, current). Plugin load mechanism has not been tested, and this prevents any functionality which actually addresses the FPGAs.
HwTile has been tested standalone, compiles with the MccsTile
Tango driver, but setting the simulationMode attribute does
not switch the simulation and actual drivers.

The Tile device must have two properties, TileIp and TpmCpldPort
set to identify IP and port of the physical board.
The bitfile must be accessible from the container with the
TpmTile driver