MCCS-510 Initial message queue implementation
This commit add a generic message queue implementation that is used by
mccs controller, mccs station and mccs subrack. These devices needed the
message queue because they are used by mccs controller during the call
to "On" as part of the pools within mccs controller. The station pools
have been stubbed out because they command tile and antenna devices to
transition to "On". This commit proves that the message system API
works, but no unit tests have been developed yet. This will be done as
part of MCCS-511. It was important to flush out the implementation on a
K8s deployment because we need to do things in a particular way due to
Tango. Also, for demonstration (and debugging) a special attribute has
been created for MCCS controller, station and subarray that show when a
message has been queued up, starts to execute and when it's complete.
This is very useful to show how the system is working, but also to show
if there has been a probloem with the underlying messaging system. I
proopse to leave this in, but it might need a little prune so the
attribute (a string) doesn't get too massive).