[MCCS-1387] Implement health in subrack
Add rules file for subrack.
This adds the rules that I think govern the subrack, however its fairly
complex with a lot going on, so I want to double check that people
are on the same page as me before I start writing a bunch of
tests that test the wrong things.
I also have no idea what values should be used as defaults for these
values, need to find those out somehow