MCCS-2191: Move initial config to DAQ receiver
Currently, a large number of DAQ receiver config values are hardcoded into the ska-low-mccs-spshw DaqComponentManager.
Additionally, the MccsDaqReceiver Tango device takes ReceiverInterface, ReceiverIp and ReceiverPort properties at deployment time.
All this configuration is then communicated by the Tango device to the DAQ, as part of the execution of the Initialise command.

It seems to me that if we are going to hardcode config values that the DAQ needs, we might as well hardcode them into the place where they are needed.
Similarly, since we already know the receiver interface, IP and port at deployment time, we might as well provide those values directly to the DAQ at deployment time, rather than providing them to a Tango device that then needs to communicate them to the DAQ. That way, the DAQ receiver is useful as soon as it is deployed, instead of having to wait for a Tango device to come along and configure it.

Of course, the Tango device should still be able to configure the DAQ receiver as it sees fit, but these properties should be optional. It should be fine for the Tango device to be deployed without these properties, in which case it accepts the DAQ receiver as it is, without trying to reconfigure it.

This commit:
* updates the ska-low-mccs-daq dependency to a version that knows its default config values and is configured with receiver interface, host and port at deployment time.
* updates the simulator to also know its default config values, and be configured at deploy time with receiver interface, host and ports.
* makes the receiverInterface, receiverHost and receiverPorts properties of the DAQ Tango device optional
* updates the DAQ Tango device to provide only config overrides to the DAQ receiver, instead of providing values for all possible config keys.