MCCS-2213: Deploy from telmodel
Implement into this repo the MCCS-2213 changes pioneered in the ska-low-mccs repo.
These changes mean that a helmfile environment does not provide a platform spec,
but rather specifies a telmodel source from which the platform spec should be retrieved.

This is basically just a copy-paste of what was done in ska-low-mccs.