[SKB-702] Multiple attempts for calibration coefficients
Some of the tiles were failing to acquire hardware lock
which meant the calibrations weren't being applied.
But the failure to acquire the lock wasn't being reported
in the result of the command, and it was still returning ok.

This leads the sps station to think the calibration was
successfully applied.

I've made the reporting better so that now the sps station
knows when the applycalibration command fails.

Also increased the timeout on the acquire hardware lock to
increase chance it succeeds.