THORN-140: reverting change to check fix
The reason why I removed this is becaused it caused the daq to fail to
stop, but now the data is never sent. So i'm reverting it in hope
that the other solution works for the fail to stop, but this fixes data
never sent error.

If I'm wrong then these errors are separate